create regular commit & release scripts. Adapted python for 3.5
